Migrate fjs/crypto/vdf to .f.mjs, split Sloth type into types.ts
Leaf module with no unmigrated dependents; updates the stale line
references in its iterate-combinator todo doc.
